+++ bundle/src/main/java/com/adobe/acs/commons/httpcache/engine/impl/HttpCacheEngineImpl.java
@@ -399,11 +399,10 @@ public class HttpCacheEngineImpl extends AnnotatedStandardMBean implements HttpC
 
         // Copy the cached data into the servlet output stream.
         try {
-            IOUtils.copy(cacheContent.getInputDataStream(), response.getOutputStream());
+            serveCacheContentIntoResponse(response, cacheContent);
             if (log.isDebugEnabled()) {
                 log.debug("Response delivered from cache for the url [ {} ]", request.getRequestURI());
             }
-
             return true;
         } catch (IOException e) {
             throw new HttpCacheDataStreamException("Unable to copy from cached data to the servlet output stream.");
@@ -528,8 +527,8 @@ public class HttpCacheEngineImpl extends AnnotatedStandardMBean implements HttpC
 
     }
 
-    //-------------------------<Mbean specific implementation>
 
+    //-------------------------<Mbean specific implementation>
     public HttpCacheEngineImpl() throws NotCompliantMBeanException {
         super(HttpCacheEngineMBean.class);
     }
@@ -627,5 +626,15 @@ public class HttpCacheEngineImpl extends AnnotatedStandardMBean implements HttpC
         }
 
         return tabularData;
+    }
+
+    private void serveCacheContentIntoResponse(SlingHttpServletResponse response, CacheContent cacheContent)
+            throws IOException {
+        try {
+            IOUtils.copy(cacheContent.getInputDataStream(), response.getOutputStream());
+        } catch (IllegalStateException ex) {
+            // in this case, either the writer has already been obtained or the response doesn't support getOutputStream()
+            IOUtils.copy(cacheContent.getInputDataStream(), response.getWriter(), response.getCharacterEncoding());
+        }
     }
 }